+++ .github/workflows/code-pushup.yml
@@ -27,7 +27,7 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v4
       - name: Set up Node.js
-        uses: actions/setup-node@v4
+        uses: actions/env-setup-node@v4
         with:
           node-version-file: .nvmrc
           cache: npm
