+++ README.md
@@ -226,7 +226,7 @@ Root/
 │            │        └── <package-name>/...
 │            ├── __test__/...
 │            │   └── <test-file-name>/...
-│            │        └── <it-block-setup>/...
+│            │        └── <it-block-env-setup>/...
 │            │             └── test.file.ts
 │            ├── .npmrc # local npm config configured for project specific Verdaccio registry
 │            ├── package-lock.json # skipped creation by default
