+++ examples/e2e/cli-e2e-original/vite.config.ts
@@ -19,7 +19,7 @@ export default defineConfig({
     environment: 'node',
     include: ['test/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}'],
     reporters: ['default'],
-    globalSetup: ['./setup/global-setup.ts'],
+    globalSetup: ['./env-setup/global-env-setup.ts'],
     coverage: {
       reportsDirectory: '../../../coverage/projects/cli-e2e-original',
       provider: 'v8',
