+++ examples/projects/cli/src/bin.ts
@@ -2,5 +2,5 @@
 import { hideBin } from 'yargs/helpers';
 import { cli } from './lib/cli';
 
-// bootstrap Yargs, parse arguments and execute command.
+// env-bootstrap Yargs, parse arguments and execute command.
 cli(hideBin(process.argv)).parseAsync();
