+++ projects/nx-verdaccio-env/src/executors/env-bootstrap/constants.ts
