+++ projects/nx-verdaccio-env/src/executors/env-bootstrap/schema.ts
