+++ projects/nx-verdaccio-env/src/executors/env-bootstrap/verdaccio-registry.ts
