+++ projects/nx-verdaccio-env/src/executors/env-setup/constants.ts
