+++ projects/nx-verdaccio-env/src/executors/env-setup/schema.ts
