+++ projects/nx-verdaccio-env/src/executors/pkg-install/README.md
