+++ projects/nx-verdaccio-env/src/executors/pkg-install/executor.unit-test.ts
