+++ projects/nx-verdaccio-env/src/executors/pkg-install/schema.ts
