+++ projects/nx-verdaccio-env/src/executors/pkg-publish/schema.json
