+++ projects/nx-verdaccio-env/src/plugin/targets/package.targets.ts
@@ -1,8 +1,8 @@
 import type { ProjectConfiguration, TargetConfiguration } from '@nx/devkit';
 import type { NormalizedCreateNodeOptions } from '../normalize-create-nodes-options';
 import { PACKAGE_NAME } from '../constants';
-import { EXECUTOR_PACKAGE_NPM_PUBLISH } from '../../executors/npm-publish/constants';
-import { EXECUTOR_PACKAGE_NPM_INSTALL } from '../../executors/npm-install/constants';
+import { EXECUTOR_PACKAGE_NPM_PUBLISH } from '../../executors/pkg-publish/constants';
+import { EXECUTOR_PACKAGE_NPM_INSTALL } from '../../executors/pkg-install/constants';
 
 export const TARGET_PACKAGE_NPM_INSTALL = 'pb-ve-pkg-install';
 export const TARGET_PACKAGE_NPM_PUBLISH = 'pb-ve-pkg-publish';
